fix: review thx @heygul
