+++ backend/cdb/api/templates/caf_msa_update_report.html
@@ -14,11 +14,11 @@
 		<p>
 			Le traitement du fichier CAF/MSA est terminé.
 			<br>
-			Nombre de dossier reçus: {{count}}
+			Nombre de dossier reçus : {{count}}
 			<br>
-			Nombre de dossiers mis à jour: {{count_success}}
+			Nombre de dossier mis à jour : {{count_success}}
 			<br>
-			Nombre de dossiers en erreur: {{count_error}}
+			Nombre de dossier en erreur : {{count_error}}
 		</p>
 		<p>L'équipe Carnet de bord</p>
 	</body>
